requirements.txt
torch==1.1.0
torchsummary==1.5.1
torchvision==0.3.0
numpy==1.17.0
Pillow==6.0.0
PyYAML==5.1.2
requests==2.22.0
tqdm==4.38.0
opencv-python
